+++ .eslintrc.js
@@ -1,14 +1,16 @@
 require("@jobber/eslint-config/patch-eslint-plugin-resolution.js");
 
+const packageAliases = [
+  ["@jobber/components", "./packages/components/src"],
+  ["@jobber/hooks", "./packages/hooks"]
+];
+
 module.exports = {
   extends: ["@jobber/eslint-config"],
   settings: {
     "import/resolver": {
       alias: {
-        map: [
-          ["@jobber/components", "./packages/components/src"],
-          ["@jobber/hooks", "./packages/hooks"]
-        ],
+        map: packageAliases,
         extensions: [".ts", ".tsx", ".js", ".jsx", ".json"]
       }
     }
@@ -26,5 +28,38 @@ module.exports = {
         allow: ["@jobber/components/*", "gatsby-theme-docz/**", "lodash/*"]
       }
     ]
-  }
+  },
+  overrides: [
+    {
+      /**
+       * Disabling default exports for a few files as docz requires
+       * some files to have default exports.
+       */
+      files: [
+        "./packages/docz-tools/src/gatsby-theme-docz/theme/index.ts",
+        "./packages/docz-tools/src/gatsby-theme-docz/components/index.ts"
+      ],
+      rules: {
+        "import/no-default-export": "off"
+      }
+    },
+    {
+      /**
+       * Adds an alias that is only available in the docz-tools
+       * package. We do not want this available anywhere else.
+       */
+      files: ["./packages/docz-tools/**/*"],
+      settings: {
+        "import/resolver": {
+          alias: {
+            map: [
+              ...packageAliases,
+              ["~theme", "./packages/docz-tools/src/gatsby-theme-docz/theme"]
+            ],
+            extensions: [".ts", ".tsx", ".js", ".jsx", ".json"]
+          }
+        }
+      }
+    }
+  ]
 };