+++ packages/components/src/Text/Text.tsx
@@ -2,6 +2,8 @@ import React, { PropsWithChildren } from "react";
 import { Typography, TypographyOptions } from "../Typography";
 
 interface TextProps {
+  maxLines?: "single" | "small" | "base" | "large" | "larger" | "unlimited";
+
   readonly variation?:
     | "default"
     | "subdued"
@@ -19,6 +21,7 @@ export function Text({
   variation = "default",
   size = "base",
   children,
+  maxLines = "unlimited",
 }: PropsWithChildren<TextProps>) {
   const textColors = {
     default: "greyBlueDark",
@@ -29,8 +32,21 @@ export function Text({
     info: "lightBlue",
   };
 
+  const maxLineToNumber = {
+    single: 1,
+    small: 2,
+    base: 4,
+    large: 8,
+    larger: 16,
+    unlimited: undefined,
+  };
+
   return (
-    <Typography textColor={textColors[variation] as TextColor} size={size}>
+    <Typography
+      textColor={textColors[variation] as TextColor}
+      size={size}
+      numberOfLines={maxLineToNumber[maxLines]}
+    >
       {children}
     </Typography>
   );