+++ packages/docz-tools/index.js
@@ -1,0 +1,5 @@
+/**
+ * A main entry point is required by a `gatsy plugin`.
+ * This file can also be used to export components for the theme
+ * when the time comes.
+ */