+++ packages/docz-tools/src/gatsby-theme-docz/components/Actions/styles.ts
@@ -1,0 +1,8 @@
+export const actions = {
+  display: "flex",
+  gap: "small",
+  position: "fixed",
+  top: "large",
+  right: "large",
+  zIndex: 9999,
+} as const;