+++ packages/docz-tools/src/gatsby-theme-docz/components/Logo/styles.ts
@@ -1,0 +1,22 @@
+export const logo = {
+  display: "block",
+  bg: "blueDark",
+  padding: "large",
+  textDecoration: "none",
+};
+
+export const text = {
+  background: `-webkit-linear-gradient(var(--color-yellow--lighter),var(--color-yellow))`,
+  WebkitBackgroundClip: "text",
+  WebkitTextFillColor: "transparent",
+  textTransform: "uppercase",
+  fontFamily: "heading",
+  letterSpacing: "0.05em",
+  fontSize: "larger",
+  padding: 0,
+};
+
+export const image = {
+  maxHeight: "largest",
+  padding: 0,
+};