+++ packages/docz-tools/src/gatsby-theme-docz/theme/space.ts
@@ -1,0 +1,11 @@
+export const space = {
+  minuscule: "var(--space-minuscule)",
+  smallest: "var(--space-smallest)",
+  smaller: "var(--space-smaller)",
+  small: "var(--space-small)",
+  base: "var(--space-base)",
+  large: "var(--space-large)",
+  larger: "var(--space-larger)",
+  largest: "var(--space-largest)",
+  extravagant: "var(--space-extravagant)",
+};