+++ packages/stylelint-config/CHANGELOG.md
@@ -8,6 +8,17 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/GetJobber/atlantis/compare/@jobber/stylelint-config@0.1.0...@jobber/stylelint-config@0.3.0) (2021-02-01)
+
+
+### Features
+
+* Readily Release Releases ([#189](https://github.com/GetJobber/atlantis/issues/189)) ([f12f518](https://github.com/GetJobber/atlantis/commit/f12f518443d5c4640d4d6cb95dc6b199b404bf8d))
+
+
+
+
+
 # [0.2.0](https://github.com/GetJobber/atlantis/compare/@jobber/stylelint-config@0.1.0...@jobber/stylelint-config@0.2.0) (2020-01-23)